Make NavMeshCacheItem consistent
Move all logic related to this type into its methods.
